requirements.txt
# Core Framework
Flask>=2.3.0,<3.0.0
Flask-RESTful>=0.3.10,<1.0.0
Werkzeug>=2.3.0,<3.0.0

# Database
pymongo>=4.6.0,<5.0.0

# Telegram Bot
python-telegram-bot>=20.0,<21.0

# Computer Vision & AWS
opencv-python-headless>=4.8.0,<5.0.0
boto3>=1.28.0,<2.0.0
numpy>=1.24.0,<2.0.0

# Utilities
geopy>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
Pillow>=10.0.0,<11.0.0
requests>=2.31.0,<3.0.0

# Scheduling
APScheduler>=3.10.0,<4.0.0
pytz>=2023.3

# Google Drive (optional)
google-api-python-client>=2.100.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.0.0,<2.0.0